requirements.txt
attrs==23.1.0
async-timeout==4.0.3
anthropic==0.10.0
huggingface-hub==0.19.4
ipdb==0.13.13
langchain==0.0.341
langchain-core==0.0.6
langsmith==0.0.67
matplotlib==3.8.3
numpy==1.26.2
openai==1.3.5
packaging==23.2
pandas==2.2.0
pillow==10.2.0
python-dotenv==1.0.0
pytz==2024.1
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
safetensors==0.4.2
seaborn==0.13.2
tiktoken==0.5.1
tokenizers==0.15.0
torch==2.2.1
torchaudio==2.2.1
torchvision==0.17.1
tqdm==4.66.1
transformers==4.38.1
triton==2.2.0
zhipuai==2.0.1
jsonpointer==2.4
